+++ pyo3-ffi/src/lib.rs
@@ -447,6 +447,7 @@ pub use self::pystate::*;
 pub use self::pystrtod::*;
 pub use self::pythonrun::*;
 pub use self::rangeobject::*;
+pub use self::refcount::*;
 pub use self::setobject::*;
 pub use self::sliceobject::*;
 pub use self::structseq::*;
@@ -539,6 +540,7 @@ mod pystrtod;
 // skipped pythread.h
 // skipped pytime.h
 mod rangeobject;
+mod refcount;
 mod setobject;
 mod sliceobject;
 mod structseq;
